Рефакторинг. Помимо рефакторинга для трех однотипных запросов к репозиторию/бд внедрил асинхрон с кастомным Executor под эту задачу.
Среднее время "старого" метода = 16ms, "нового" = 9ms.
