+++ orchestrator_service/src/main/java/com/override/orchestrator_service/config/DiffWidgetAsyncConfig.java
@@ -1,0 +1,19 @@
+package com.override.orchestrator_service.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.task.TaskExecutor;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+
+@Configuration
+public class DiffWidgetAsyncConfig {
+
+    public TaskExecutor widgetAsyncExecutor() {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+        executor.setCorePoolSize(3);
+        executor.setMaxPoolSize(3);
+        executor.setQueueCapacity(20);
+        executor.setThreadNamePrefix("diff-widget-async-thread-");
+        executor.initialize();
+        return executor;
+    }
+}
